alicorn/requirements.txt
PyYAML==3.10
irc==8.3
requests==1.2.3
